Create GMS and HMS Loction controller
   * Separate location controller by platform
