+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/HmsMessageServiceOneSignal.java
@@ -1,9 +1,6 @@
 package com.onesignal;
 
-import android.util.Log;
-
 import com.huawei.hms.push.HmsMessageService;
-import com.huawei.hms.push.RemoteMessage;
 
 public class HmsMessageServiceOneSignal extends HmsMessageService {
