+++ OneSignalSDK/unittest/src/test/java/com/onesignal/ShadowHMSLocationUpdateListener.java
@@ -1,7 +1,5 @@
 package com.onesignal;
 
-import android.location.Location;
-
 import com.huawei.hms.location.HWLocation;
 import com.huawei.hms.location.LocationResult;
 
@@ -10,22 +8,13 @@ import org.robolectric.annotation.Implements;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Copyright 2017 OneSignal
- * Created by alamgir on 9/22/17.
- */
-
-@Implements(LocationController.LocationUpdateListener.class)
-public class ShadowLocationUpdateListener {
-
-    public static void provideFakeLocation(Location location) {
-        LocationController.locationUpdateListener.onLocationChanged(location);
-    }
+@Implements(HMSLocationController.LocationUpdateListener.class)
+public class ShadowHMSLocationUpdateListener {
 
     public static void provideFakeLocation_Huawei(HWLocation location) {
         List<HWLocation> locations = new ArrayList<>();
         locations.add(location);
         LocationResult locationResult = LocationResult.create(locations);
-        LocationController.locationUpdateListener.onLocationResult(locationResult);
+        HMSLocationController.locationUpdateListener.onLocationResult(locationResult);
     }
 }
