+++ OneSignalSDK/unittest/src/test/java/com/onesignal/ShadowHmsInstanceId.java
@@ -1,7 +1,6 @@
 package com.onesignal;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import com.huawei.hms.aaid.HmsInstanceId;
