+++ OneSignalSDK/unittest/src/test/java/com/test/onesignal/OutcomeEventIntegrationTests.java
@@ -17,7 +17,7 @@ import com.onesignal.ShadowAdvertisingIdProviderGPS;
 import com.onesignal.ShadowCustomTabsClient;
 import com.onesignal.ShadowCustomTabsSession;
 import com.onesignal.ShadowJobService;
-import com.onesignal.ShadowLocationController;
+import com.onesignal.ShadowGMSLocationController;
 import com.onesignal.ShadowNotificationManagerCompat;
 import com.onesignal.ShadowOSUtils;
 import com.onesignal.ShadowOneSignalRestClient;
@@ -75,7 +75,7 @@ import static junit.framework.Assert.assertTrue;
         shadows = {
                 ShadowOneSignalRestClient.class,
                 ShadowPushRegistratorGCM.class,
-                ShadowLocationController.class,
+                ShadowGMSLocationController.class,
                 ShadowOSUtils.class,
                 ShadowAdvertisingIdProviderGPS.class,
                 ShadowCustomTabsClient.class,
